+++ src/dataset/Dataset.present.js
@@ -29,7 +29,7 @@ import { ProjectsCoordinator } from "../project/shared";
 function DisplayFiles(props) {
   if (props.files === undefined) return null;
 
-  if (props.files.error) {
+  if (props.files.error !== undefined) {
     return <Card key="datasetDetails">
       <CardHeader className="align-items-baseline">
         <span className="caption align-baseline">Dataset files</span>
